Fixed master content contamination
new_content_list was intended to be totally separate from the original master. However, the tuple contained references to the original content lists, so when we added the flattened content in, it would contaminate the original master.

If the instance was used multiple times in the hierachy, it would contain multiple copies of the same flattened content, leading the memory consumption that was quadratic with number of instances.